--- .parcelrc
@@ -1,8 +1,0 @@
-{
-  "extends": "@parcel/config-default",
-  "transformers": {
-    "*.css": [
-      "parcel-transformer-css-to-string"
-    ]
-  }
-}
