+++ demo/electron-tabs.html
@@ -15,7 +15,7 @@
   </style>
 </tab-group>
 
-<script src="../dist/electron-tabs.sortable.js"></script>
+<script src="../dist/electron-tabs.js"></script>
 <script>
   const tabGroup = document.querySelector("tab-group");
   tabGroup.on("ready", () => console.info("TabGroup is ready"));
