+++ package.json
@@ -3,13 +3,11 @@
   "version": "1.0.0-dev",
   "description": "Simple tabs for Electron applications",
   "main": "dist/electron-tabs.js",
-  "sortable": "dist/electron-tabs.sortable.js",
+  "types": "dist/electron-tabs.d.ts",
+  "source": "src/index.ts",
   "targets": {
     "main": {
-      "source": "src/index.js"
-    },
-    "sortable": {
-      "source": "src/sortable.js"
+      "includeNodeModules": true
     }
   },
   "repository": {
@@ -30,11 +28,15 @@
   "author": "brrd",
   "license": "MIT",
   "devDependencies": {
+    "@parcel/packager-ts": "^2.5.0",
+    "@parcel/transformer-inline-string": "^2.5.0",
+    "@parcel/transformer-typescript-types": "^2.5.0",
+    "@types/sortablejs": "^1.13.0",
     "cssnano": "^5.1.9",
     "electron": "^17.1.2",
     "parcel": "^2.5.0",
-    "parcel-transformer-css-to-string": "^0.9.1",
     "postcss": "^8.4.14",
-    "sortablejs": "^1.15.0"
+    "sortablejs": "^1.15.0",
+    "typescript": "^4.7.2"
   }
 }
