--- src/sortable.js
@@ -1,4 +1,0 @@
-import Sortable from "sortablejs";
-import "./index.js";
-
-window.Sortable = Sortable;
